pyproject.toml
[build-system]
requires = ["setuptools>=77.0"]
build-backend = "setuptools.build_meta"

[project]
name = "LeapAI-Assistant"
version = "0.1.0"
description = "AI-native onboarding assistant for openSUSE Leap"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
authors = [
    { name = "Yasharth Panwar", email = "panwaryasharth203@gmail.com" },
]
dependencies = [
    # Core Orchestration
    "textual>=8.0.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.0",
    "aiohttp>=3.11.0",
    "aiofiles>=25.1.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",

    # RAG Backends
    "qdrant-client>=1.17.0",
    "chromadb",

    # Ingestion Pipeline
    "crawl4ai>=0.8.5",
    "feedparser>=6.0.11",
    "html2text",
    "beautifulsoup4>=4.14.0",
    "lxml>=5.0.0",
    "fake-useragent>=2.2.0",

    # Embeddings & AI/LLM
    "sentence-transformers",
    "openai>=2.0.0",
    "unclecode-litellm>=1.81.0",
    "tiktoken>=0.12.0",

    # NLP & Data/Geo
    "nltk>=3.9.0",
    "rank-bm25>=0.2.2",
    "trimesh>=4.11.0",
    "alphashape>=1.3.0",
    "shapely>=2.0.0",

    # Utilities
    "rich>=14.0.0",
    "psutil>=7.0.0",
    "python-discovery>=1.2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.0",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "types-requests>=2.32.0.20250602",
    "types-PyYAML>=6.0.12.20250516",
]

[project.scripts]
suse-ai = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
asyncio_default_test_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
pythonpath = ["."]
